apps/api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
pydantic-settings==2.6.1
httpx==0.28.1
apscheduler==3.10.4
stripe==12.2.0
